third_party/WebKit/LayoutTests/http/tests/security/popup-allowed-by-sandbox-when-allowed-expected.txt

CONSOLE MESSAGE: line 5: PASS
To run this test outside of DumpRenderTree, please disable your popup blocker!
